# OrigenX/BackEnd/openapi.yaml
openapi: 3.0.3
info:
  title: The Artisanal Connection API
  description: |
    API REST para The Artisanal Connection. Conecta productores de cafe colombiano con consumidores finales.

    ## Autenticacion
    La mayoria de endpoints requieren un Firebase ID Token en el header:
    `Authorization: Bearer <firebase_id_token>`

    ## Roles
    - **CONSUMER**: consumidor final que navega el catalogo y realiza pedidos.
    - **PRODUCER**: productor de cafe que gestiona productos y ventas.
    - **ADMIN**: administrador de la plataforma.

    ## Formato de errores
    Todos los errores siguen el formato uniforme:
    ```json
    {
      "error": {
        "code": "VALIDATION_ERROR",
        "message": "Descripcion del error",
        "fields": [{"field": "email", "message": "El correo ya esta registrado"}]
      }
    }
    ```
  version: 0.1.0
  contact:
    name: The Artisanal Connection
servers:
  - url: http://localhost:8000
    description: Servidor de desarrollo local
  - url: https://api.marketplacecafetero.com
    description: Servidor de produccion
...
security:
  - BearerAuth: []
paths:
...
  /api/auth/register:
    post:
      tags: [auth]
      summary: Registro de usuario
      description: Crea una nueva cuenta de usuario. Retorna HTTP 409 si el correo ya existe. Requerimientos 1.1-1.5.
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
            example:
              name: Juan Perez
              email: juan@example.com
              password: miContrasena123
              role: CONSUMER
      responses:
        "201":
          description: Usuario registrado exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
              example:
                message: Usuario registrado exitosamente.
                data:
                  id: abc123uid
                  name: Juan Perez
                  email: juan@example.com
                  role: CONSUMER
                  created_at: "2024-01-15T10:30:00"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/auth/login:
    post:
      tags: [auth]
      summary: Inicio de sesion
      description: Verifica el Firebase ID Token del usuario. El cliente debe autenticarse con Firebase Client SDK y enviar el ID Token en el campo password. Retorna HTTP 429 tras 5 intentos fallidos. Requerimientos 2.1-2.2.
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              email: juan@example.com
              password: eyJhbGciOiJSUzI1NiJ9...
      responses:
        "200":
          description: Sesion iniciada exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "429":
          $ref: "#/components/responses/RateLimitError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/auth/forgot-password:
    post:
      tags: [auth]
      summary: Solicitar recuperacion de contrasena
      description: Envia un email de restablecimiento. Siempre retorna el mismo mensaje sin revelar si el correo existe. Requerimientos 2.4-2.5.
      operationId: forgotPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          description: Solicitud procesada
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Si el correo esta registrado, recibiras un enlace de restablecimiento en los proximos minutos.
        "500":
          $ref: "#/components/responses/InternalError"
  /api/auth/reset-password:
    post:
      tags: [auth]
      summary: Restablecer contrasena
      description: Restablece la contrasena usando el token enviado por email. Requerimiento 2.4.
      operationId: resetPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Contrasena restablecida exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Contrasena restablecida exitosamente.
        "400":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/addresses:
    get:
      tags: [addresses]
      summary: Listar direcciones de envio
      description: Retorna todas las direcciones de envio del consumidor autenticado. Requiere rol CONSUMER. Requerimiento 18.4.
      operationId: listAddresses
      responses:
        "200":
          description: Direcciones del consumidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AddressResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [addresses]
      summary: Registrar direccion de envio
      description: "Registra una nueva direccion de envio. Campos obligatorios: street, city, department. Limite de 5 direcciones por usuario. Requiere rol CONSUMER. Requerimientos 18.1-18.3."
      operationId: createAddress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressCreate"
            example:
              street: "Calle 10 No 5-20"
              city: Bogota
              department: Cundinamarca
              postal_code: "110111"
      responses:
        "201":
          description: Direccion registrada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Direccion registrada exitosamente.
                  data:
                    $ref: "#/components/schemas/AddressResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/addresses/{address_id}:
    delete:
      tags: [addresses]
      summary: Eliminar direccion de envio
      description: Elimina una direccion de envio del consumidor. Requiere rol CONSUMER.
      operationId: deleteAddress
      parameters:
        - $ref: "#/components/parameters/AddressId"
      responses:
        "200":
          description: Direccion eliminada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Direccion eliminada exitosamente.
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID Token
      description: "Firebase ID Token obtenido mediante Firebase Client SDK. Incluir en el header Authorization: Bearer <token>"
  parameters:
...
    AddressId:
      name: address_id
      in: path
      required: true
      description: Identificador unico de la direccion
      schema:
        type: string
        example: addr789pqr
  schemas:
    ErrorField:
      type: object
      properties:
        field:
          type: string
          example: email
        message:
          type: string
          example: El correo ya esta registrado
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Descripcion del error
        fields:
          type: array
          items:
            $ref: "#/components/schemas/ErrorField"
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
    UserCreate:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: Juan Perez
        email:
          type: string
          format: email
          example: juan@example.com
        password:
          type: string
          minLength: 8
          description: Minimo 8 caracteres
          example: miContrasena123
        role:
          type: string
          enum: [CONSUMER, PRODUCER]
          default: CONSUMER
          description: Rol del usuario. ADMIN no puede registrarse via API.
    RegisterResponse:
      type: object
      properties:
        message:
          type: string
          example: Usuario registrado exitosamente.
        data:
          type: object
          properties:
            id:
              type: string
              example: abc123uid
            name:
              type: string
              example: Juan Perez
            email:
              type: string
              format: email
              example: juan@example.com
            role:
              type: string
              enum: [CONSUMER, PRODUCER, ADMIN]
              example: CONSUMER
            created_at:
              type: string
              format: date-time
              example: "2024-01-15T10:30:00"
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: juan@example.com
        password:
          type: string
          description: Firebase ID Token obtenido del cliente Firebase SDK
          example: eyJhbGciOiJSUzI1NiJ9...
    LoginResponse:
      type: object
      properties:
        message:
          type: string
          example: Sesion iniciada exitosamente.
        data:
          type: object
          properties:
            uid:
              type: string
              example: abc123uid
            email:
              type: string
              format: email
              example: juan@example.com
            role:
              type: string
              enum: [CONSUMER, PRODUCER, ADMIN]
              example: CONSUMER
            name:
              type: string
              example: Juan Perez
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          example: juan@example.com
    ResetPasswordRequest:
      type: object
      required: [token, new_password]
      properties:
        token:
          type: string
          description: Token de restablecimiento enviado por email
          example: oobCode123
        new_password:
          type: string
          minLength: 8
          description: Nueva contrasena (minimo 8 caracteres)
          example: nuevaContrasena456
...
    AddressCreate:
      type: object
      required: [street, city, department]
      properties:
        street:
          type: string
          minLength: 1
          description: Calle (obligatorio)
          example: Calle 10 # 5-20
        city:
          type: string
          minLength: 1
          description: Ciudad (obligatorio)
          example: Bogota
        department:
          type: string
          minLength: 1
          description: Departamento (obligatorio)
          example: Cundinamarca
        postal_code:
          type: string
          nullable: true
          example: "110111"
    AddressResponse:
      type: object
      properties:
        id:
          type: string
          example: addr789pqr
        user_id:
          type: string
          example: abc123uid
        street:
          type: string
          example: Calle 10 # 5-20
        city:
          type: string
          example: Bogota
        department:
          type: string
          example: Cundinamarca
        postal_code:
          type: string
          nullable: true
          example: "110111"
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00"
  responses:
    ValidationError:
      description: Datos de entrada invalidos (HTTP 400)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: VALIDATION_ERROR
              message: Los datos de entrada son invalidos.
              fields:
                - field: email
                  message: El correo electronico no tiene un formato valido.
                - field: password
                  message: La contrasena debe tener al menos 8 caracteres.
    UnauthorizedError:
      description: No autenticado - JWT ausente o invalido (HTTP 401)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: UNAUTHORIZED
              message: Token de autenticacion ausente o invalido.
    ForbiddenError:
      description: No autorizado - rol insuficiente o recurso ajeno (HTTP 403)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: FORBIDDEN
              message: No tienes permiso para realizar esta accion.
    NotFoundError:
      description: Recurso no encontrado (HTTP 404)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: NOT_FOUND
              message: El recurso solicitado no fue encontrado.
    ConflictError:
      description: Conflicto - recurso ya existe (HTTP 409)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: EMAIL_ALREADY_EXISTS
              message: El correo electronico ya esta registrado.
              fields:
                - field: email
                  message: El correo ya esta registrado.
...
    RateLimitError:
      description: Demasiadas solicitudes - rate limiting (HTTP 429)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: RATE_LIMIT_EXCEEDED
              message: Demasiados intentos fallidos. Intenta de nuevo en 15 minutos.
    InternalError:
      description: Error interno del servidor (HTTP 500)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: INTERNAL_ERROR
              message: Error interno del servidor. Por favor intenta de nuevo.
...